toolchain: use size_t in stdio surface

+++ toolchain/include/stdio.h
@@ -2,6 +2,7 @@
 #define LNP64_STDIO_H
 
 #include <stdarg.h>
+#include <stddef.h>
 
 typedef struct __lnp64_file FILE;
 
@@ -36,15 +37,13 @@ long ftello(FILE *stream);
 int fscanf(FILE *stream, const char *format, ...);
 int vfprintf(FILE *stream, const char *format, va_list ap);
 int sprintf(char *str, const char *format, ...);
-int vsnprintf(char *str, unsigned long size, const char *format, va_list ap);
-int snprintf(char *str, unsigned long size, const char *format, ...);
+int vsnprintf(char *str, size_t size, const char *format, va_list ap);
+int snprintf(char *str, size_t size, const char *format, ...);
 long getline(char **lineptr, unsigned long *n, FILE *stream);
-unsigned long fread(void *ptr, unsigned long size, unsigned long count,
-                    FILE *stream);
-unsigned long fwrite(const void *ptr, unsigned long size, unsigned long count,
-                     FILE *stream);
+size_t fread(void *ptr, size_t size, size_t count, FILE *stream);
+size_t fwrite(const void *ptr, size_t size, size_t count, FILE *stream);
 char *fgets(char *str, int count, FILE *stream);
-FILE *fmemopen(void *buf, unsigned long size, const char *mode);
+FILE *fmemopen(void *buf, size_t size, const char *mode);
 FILE *fdopen(int fd, const char *mode);
 FILE *fopen(const char *path, const char *mode);
 FILE *tmpfile(void);
